requirements.txt
tqdm
PyMuPDF>=1.23.8   
rank-bm25>=0.2.2
chromadb>=0.5.0
sentence-transformers>=2.6.1
huggingface-hub>=0.23.0
llama-cpp-python[cuda12]==0.3.16
regex